requirements.txt
boto3
edge-tts
keyboard
numpy
openai-whisper
pillow
pydub
python-osc
pyspellchecker
sounddevice
soundfile
torch==2.5.1
webrtcvad
